add feedback from lpc

+++ lpc_pi7/pi7/src/pi7/comunicacao_pic/comunicacao_pic.h
@@ -6,9 +6,9 @@
  */
 
 typedef struct {
-	float setPoint1;
-	float setPoint2;
-	float setPoint3;
+	int setPoint1;
+	int setPoint2;
+	int setPoint3;
 } pic_Data;
 
 extern void pic_sendToPIC(pic_Data data);
